Completed and tested pin_db

+++ Backend/SpringBackend/src/main/java/com/infoFootball/SpringBackend/Pins/Pin.java
@@ -15,8 +15,8 @@ public class Pin {
 
     private String username;
 
-    private Long latitude;
-    private Long longitude;
+    private Double latitude;
+    private Double longitude;
 
     private String pinName;
 
@@ -28,7 +28,7 @@ public class Pin {
         this.pinName = null;
     }
 
-    public Pin(String username, Long latitude, Long longitude, String pinName){
+    public Pin(String username, Double latitude, Double longitude, String pinName){
         this.username = username;
         this.latitude = latitude;
         this.longitude = longitude;
@@ -43,19 +43,19 @@ public class Pin {
         this.username = username;
     }
 
-    public long getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public long getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
